memperbaiki font size serta menambahkan content awal buka pada dashboard
