+++ src/templates/index.html
@@ -32,5 +32,5 @@
     <footer>
         <p>Copyright @Cari Paket 2021</p>
     </footer>
-</body>
+</body>                   
 </html>
